+++ subject-list.php
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <?php require 'db_connection.php';?>
+    <?php require 'db_connection.php'; ?>
     <meta charset="UTF-8">
     <title>Title</title>
     <style>
@@ -14,28 +14,42 @@
 <body>
 
 <table width="100%">
+    <thead>
     <tr>
         <th>Name</th>
         <th>Semester</th>
-        <th>ECDL</th>
-        <th>Edit</th>
-        <th>Delete</th>
+        <th>ECDL credits</th>
+        <th>Description</th>
+        <?php
+        if (isset($_SESSION["username"])) {
+            if ($_SESSION["is_admin"] == 1) {
+                echo "<th>Edit</th>";
+                echo "<th>Delete</th>";
+            }
+        }
+        ?>
     </tr>
+    </thead>
+    <tbody>
     <?php
-    $result = mysqli_query($conn,"SELECT * FROM subject");
-
-    while($row = mysqli_fetch_array($result))
-    {
+    $result = mysqli_query($conn, "SELECT * FROM subject");
+    while ($row = mysqli_fetch_array($result)) {
         echo "<tr>";
         echo "<td>" . $row['name'] . "</td>";
         echo "<td>" . $row['semester'] . "</td>";
         echo "<td>" . $row['ecdl_credits'] . "</td>";
-        echo "<td><a href='edit-subject.php?id=".$row['id']."'>Edit</a></td>";
-        echo "<td><a href='delete-subject.php?id=".$row['id']."'>Delete</a></td>";
+        echo "<td>" . $row['description'] . "</td>";
+        if (isset($_SESSION["username"])) {
+            if ($_SESSION["is_admin"] == 1) {
+                echo "<td><a href='edit-subject.php?id=" . $row['id'] . "'>Edit</a></td>";
+                echo "<td><a href='delete-subject.php?id=" . $row['id'] . "'>Delete</a></td>";
+            }
+        }
         echo "</tr>";
     }
     mysqli_close($conn);
     ?>
+    </tbody>
 </table>
 
 </body>
